to-do-app/backend/todo-app/pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = ""
authors = ["Alla-ud-din <allauddin.aliahmad9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlmodel = "^0.0.18"
psycopg = {extras = ["binary"], version = "^3.1.18"}
pytest = "^8.2.1"
httpx = "^0.27.0"
python-multipart = "^0.0.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
